requirements.txt
# Core dependencies
langchain==0.3.7
langchain-openai==0.2.6
langgraph==0.2.34
langsmith==0.1.137
openai>=1.54.0,<2.0.0

# Twitter scraping
twikit==2.3.3
requests==2.32.3
httpx==0.27.2

# Video processing
yt-dlp==2024.9.27
opencv-python==4.10.0.84
ffmpeg-python==0.2.0
pillow==10.4.0

# Data processing
pandas==2.2.3
numpy>=1.21,<2
pydantic==2.9.2

# CLI and utilities
click==8.1.7
rich==13.8.1
tqdm==4.66.5
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0

# Development
black==24.8.0
flake8==7.1.1
mypy==1.11.2